vimailserver/vidb/requirements.txt
adal==1.2.2
asn1crypto==0.24.0
azure-common==1.1.23
azure-keyvault==1.1.0
azure-nspkg==3.0.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
idna==2.8
isodate==0.6.0
MarkupSafe==1.1.1
msrest==0.6.8
msrestazure==0.6.1
oauthlib==3.1.0
passlib==1.7.1
pycparser==2.19
PyJWT==1.7.1
pymssql==2.1.4
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
